Split console formatter on run time vs. memory usage
I was having some issues keeping things straight with everything in one
file for the console formatter, especially since there was so much that
was _similar_ but subtly different, so now I've pulled that apart into
three modules so it'll be easier to keep stuff straight.
