+++ lib/benchee/conversion.ex
@@ -2,14 +2,15 @@ defmodule Benchee.Conversion do
   @moduledoc """
   Integration of the conversion of multiple units with benchee.
 
-  Can be used by plugins to use benche unit scaling logic.
+  Can be used by plugins to use benchee unit scaling logic.
   """
 
   alias Benchee.Conversion.{Duration, Count, Memory}
 
   @doc """
-  Takes scenarios and a given scaling_strategy, returns the best unit for
-  run_time and ips, according to the scaling_strategy, in a map.
+  Takes scenarios and a given scaling_strategy, returns the best units for the
+  given scaling strategy. The return value changes based on whether you want
+  units for run time or memory usage.
 
   The units can then be passed on to the appropriate `format` calls to format
   the output of arbitrary values with the right unit.
@@ -18,7 +19,7 @@ defmodule Benchee.Conversion do
 
       iex> statistics = %Benchee.Statistics{average: 1000.0, ips: 1000.0}
       iex> scenario = %Benchee.Benchmark.Scenario{run_time_statistics: statistics}
-      iex> Benchee.Conversion.units([scenario], :best)
+      iex> Benchee.Conversion.units([scenario], :best, :run_time)
       %{
         ips:      %Benchee.Conversion.Unit{
                     label: "K",
@@ -34,9 +35,7 @@ defmodule Benchee.Conversion do
                   }
       }
   """
-  def units(scenarios, scaling_strategy) do
-    # Produces a map like
-    #   %{run_time: [12345, 15431, 13222], ips: [1, 2, 3]}
+  def units(scenarios, scaling_strategy, :run_time) do
     measurements =
       scenarios
       |> Enum.flat_map(fn scenario ->
@@ -51,8 +50,6 @@ defmodule Benchee.Conversion do
   end
 
   def units(scenarios, scaling_strategy, :memory) do
-    # Produces a map like
-    #   %{run_time: [12345, 15431, 13222], ips: [1, 2, 3]}
     measurements =
       scenarios
       |> Enum.flat_map(fn scenario ->
@@ -60,11 +57,16 @@ defmodule Benchee.Conversion do
       end)
       |> Enum.group_by(fn {stat_name, _} -> stat_name end, fn {_, value} -> value end)
 
-    ninety_ninth = Enum.map(measurements.percentiles, fn %{99 => percentile} -> percentile end)
+    %{memory: Memory.best(measurements.average, strategy: scaling_strategy)}
+  end
 
-    %{
-      memory: Memory.best(measurements.average, strategy: scaling_strategy),
-      ninety_ninth: Memory.best(ninety_ninth, strategy: scaling_strategy)
-    }
+  @deprecated "0.13"
+  @doc """
+  This is the old way of calling this function, which assumed that we were only
+  dealing with run times and not memory units. After we have some time to update
+  the formatters then we can remove this, but for now we'll just deprecate it.
+  """
+  def units(scenarios, scaling_strategy) do
+    units(scenarios, scaling_strategy, :run_time)
   end
 end
