+++ lib/benchee/formatters/console.ex
@@ -6,29 +6,8 @@ defmodule Benchee.Formatters.Console do
 
   use Benchee.Formatter
 
-  alias Benchee.{
-    Statistics,
-    Suite,
-    Benchmark.Scenario,
-    Configuration,
-    Conversion
-  }
-
-  alias Benchee.Conversion.{Count, Duration, Unit, DeviationPercent}
-
-  @type unit_per_statistic :: %{atom => Unit.t()}
-
-  # Length of column header
-  @default_label_width 4
-  @ips_width 13
-  @average_width 15
-  @deviation_width 11
-  @median_width 15
-  @percentile_width 15
-  @minimum_width 15
-  @maximum_width 15
-  @sample_size_width 15
-  @mode_width 25
+  alias Benchee.{Statistics, Suite, Configuration}
+  alias Benchee.Formatters.Console.{Memory, RunTime}
 
   @doc """
   Formats the benchmark statistics to a report suitable for output on the CLI.
@@ -58,7 +37,8 @@ defmodule Benchee.Formatters.Console do
   ...>     formatter_options: %{
   ...>       console: %{comparison: false, extended_statistics: false}
   ...>     },
-  ...>     unit_scaling: :best
+  ...>     unit_scaling: :best,
+  ...>     measure_memory: false
   ...>   }
   ...> }
   iex> Benchee.Formatters.Console.format(suite)
@@ -71,23 +51,16 @@ defmodule Benchee.Formatters.Console do
   """
   @spec format(Suite.t()) :: [any]
   def format(%Suite{scenarios: scenarios, configuration: config}) do
-    measure_memory = config.measure_memory
+    %{measure_memory: measure_memory} = config
+
     config = console_configuration(config)
 
     scenarios
     |> Enum.group_by(fn scenario -> scenario.input_name end)
     |> Enum.map(fn {input, scenarios} ->
-      if measure_memory do
-        [
-          input_header(input),
-          "\nRun time statistics:\n",
-          format_scenarios(scenarios, config),
-          "\nMemory usage statistics:\n",
-          format_scenarios_for_memory(scenarios, config)
-        ]
-      else
-        [input_header(input) | format_scenarios(scenarios, config)]
-      end
+      scenarios
+      |> Statistics.sort()
+      |> generate_output(config, input, measure_memory)
     end)
   end
 
@@ -101,12 +74,14 @@ defmodule Benchee.Formatters.Console do
     _ -> {:error, "Unknown Error"}
   end
 
-  defp console_configuration(%Configuration{
-         formatter_options: %{console: config},
-         unit_scaling: scaling_strategy
-       }) do
-    if Map.has_key?(config, :unit_scaling), do: warn_unit_scaling()
-    Map.put(config, :unit_scaling, scaling_strategy)
+  defp console_configuration(config) do
+    %Configuration{
+      formatter_options: %{console: console_config},
+      unit_scaling: scaling_strategy
+    } = config
+
+    if Map.has_key?(console_config, :unit_scaling), do: warn_unit_scaling()
+    Map.put(console_config, :unit_scaling, scaling_strategy)
   end
 
   defp warn_unit_scaling do
@@ -115,470 +90,24 @@ defmodule Benchee.Formatters.Console do
     )
   end
 
+  defp generate_output(scenarios, config, input, measure_memory)
+
+  defp generate_output(scenarios, config, input, true) do
+    [
+      input_header(input) |
+      RunTime.format_scenarios(scenarios, config) ++
+      Memory.format_scenarios(scenarios, config)
+    ]
+  end
+
+  defp generate_output(scenarios, config, input, false) do
+    [
+      input_header(input) |
+      RunTime.format_scenarios(scenarios, config)
+    ]
+  end
+
   @no_input_marker Benchee.Benchmark.no_input()
-  defp input_header(input) do
-    case input do
-      @no_input_marker -> ""
-      _ -> "\n##### With input #{input} #####"
-    end
-  end
-
-  @doc """
-  Formats the job statistics to a report suitable for output on the CLI.
-
-  ## Examples
-
-  ```
-  iex> scenarios = [
-  ...>   %Benchee.Benchmark.Scenario{
-  ...>     name: "My Job", run_time_statistics: %Benchee.Statistics{
-  ...>       average: 200.0, ips: 5000.0,std_dev_ratio: 0.1, median: 190.0, percentiles: %{99 => 300.1},
-  ...>       minimum: 100.1, maximum: 200.2, sample_size: 10_101, mode: 333.2
-  ...>     }
-  ...>   },
-  ...>   %Benchee.Benchmark.Scenario{
-  ...>     name: "Job 2", run_time_statistics: %Benchee.Statistics{
-  ...>       average: 400.0, ips: 2500.0, std_dev_ratio: 0.2, median: 390.0, percentiles: %{99 => 500.1},
-  ...>       minimum: 200.2, maximum: 400.4, sample_size: 20_202, mode: [612.3, 554.1]
-  ...>     }
-  ...>   }
-  ...> ]
-  iex> configuration = %{comparison: false, unit_scaling: :best, extended_statistics: true}
-  iex> Benchee.Formatters.Console.format_scenarios(scenarios, configuration)
-  ["\nName             ips        average  deviation         median         99th %\n",
-  "My Job           5 K         200 μs    ±10.00%         190 μs      300.10 μs\n",
-  "Job 2         2.50 K         400 μs    ±20.00%         390 μs      500.10 μs\n",
-  "\nExtended statistics: \n",
-  "\nName           minimum        maximum    sample size                     mode\n",
-  "My Job       100.10 μs      200.20 μs        10.10 K                333.20 μs\n",
-  "Job 2        200.20 μs      400.40 μs        20.20 K     612.30 μs, 554.10 μs\n"]
-
-  ```
-
-  """
-  @spec format_scenarios([Scenario.t()], map) :: [String.t(), ...]
-  def format_scenarios(scenarios, config) do
-    sorted_scenarios = Statistics.sort(scenarios)
-    %{unit_scaling: scaling_strategy} = config
-    units = Conversion.units(sorted_scenarios, scaling_strategy)
-    label_width = label_width(sorted_scenarios)
-
-    [
-      column_descriptors(label_width)
-      | scenario_reports(sorted_scenarios, units, label_width) ++
-          comparison_report(sorted_scenarios, units, label_width, config) ++
-          extended_statistics_report(sorted_scenarios, units, label_width, config)
-    ]
-  end
-
-  def format_scenarios_for_memory(scenarios, config) do
-    sorted_scenarios = Statistics.sort(scenarios)
-    %{unit_scaling: scaling_strategy} = config
-    units = Conversion.units(sorted_scenarios, scaling_strategy, :memory)
-    label_width = label_width(sorted_scenarios)
-
-    if all_have_deviation_of_0?(scenarios) do
-      [
-        column_descriptors_for_memory_same(label_width)
-        | scenario_reports_for_memory_same(sorted_scenarios, units, label_width) ++
-            comparison_report_for_memory(sorted_scenarios, units, label_width, config) ++
-            extended_statistics_report(sorted_scenarios, units, label_width, config) ++
-            "\n**All measurements for memory usage were the same, so no metrics could be calculated**\n"
-      ]
-    else
-      [
-        column_descriptors_for_memory(label_width)
-        | scenario_reports_for_memory(sorted_scenarios, units, label_width) ++
-            comparison_report_for_memory(sorted_scenarios, units, label_width, config) ++
-            extended_statistics_report(sorted_scenarios, units, label_width, config)
-      ]
-    end
-  end
-
-  defp all_have_deviation_of_0?(scenarios) do
-    Enum.all?(scenarios, fn scenario ->
-      scenario.memory_usage_statistics.std_dev == 0.0
-    end)
-  end
-
-  @spec extended_statistics_report([Scenario.t()], unit_per_statistic, integer, map) :: [
-          String.t()
-        ]
-  defp extended_statistics_report(_, _, _, %{extended_statistics: false}) do
-    []
-  end
-
-  defp extended_statistics_report(scenarios, units, label_width, _config) do
-    [
-      descriptor("Extended statistics"),
-      extended_column_descriptors(label_width)
-      | extended_statistics(scenarios, units, label_width)
-    ]
-  end
-
-  @spec extended_statistics([Scenario.t()], unit_per_statistic, integer) :: [String.t()]
-  defp extended_statistics(scenarios, units, label_width) do
-    Enum.map(scenarios, fn scenario ->
-      format_scenario_extended(scenario, units, label_width)
-    end)
-  end
-
-  @spec format_scenario_extended(Scenario.t(), unit_per_statistic, integer) :: String.t()
-  defp format_scenario_extended(
-         %Scenario{
-           name: name,
-           run_time_statistics: %Statistics{
-             minimum: minimum,
-             maximum: maximum,
-             sample_size: sample_size,
-             mode: mode
-           }
-         },
-         %{run_time: run_time_unit},
-         label_width
-       ) do
-    "~*s~*ts~*ts~*ts~*ts\n"
-    |> :io_lib.format([
-      -label_width,
-      name,
-      @minimum_width,
-      run_time_out(minimum, run_time_unit),
-      @maximum_width,
-      run_time_out(maximum, run_time_unit),
-      @sample_size_width,
-      Count.format(sample_size),
-      @mode_width,
-      mode_out(mode, run_time_unit)
-    ])
-    |> to_string
-  end
-
-  @spec mode_out(Statistics.mode(), Benchee.Conversion.Unit.t()) :: String.t()
-  defp mode_out(modes, _run_time_unit) when is_nil(modes) do
-    "None"
-  end
-
-  defp mode_out(modes, run_time_unit) when is_list(modes) do
-    Enum.map_join(modes, ", ", fn mode -> run_time_out(mode, run_time_unit) end)
-  end
-
-  defp mode_out(mode, run_time_unit) when is_number(mode) do
-    run_time_out(mode, run_time_unit)
-  end
-
-  @spec extended_column_descriptors(integer) :: String.t()
-  defp extended_column_descriptors(label_width) do
-    "\n~*s~*s~*s~*s~*s\n"
-    |> :io_lib.format([
-      -label_width,
-      "Name",
-      @minimum_width,
-      "minimum",
-      @maximum_width,
-      "maximum",
-      @sample_size_width,
-      "sample size",
-      @mode_width,
-      "mode"
-    ])
-    |> to_string
-  end
-
-  @spec column_descriptors(integer) :: String.t()
-  defp column_descriptors(label_width) do
-    "\n~*s~*s~*s~*s~*s~*s\n"
-    |> :io_lib.format([
-      -label_width,
-      "Name",
-      @ips_width,
-      "ips",
-      @average_width,
-      "average",
-      @deviation_width,
-      "deviation",
-      @median_width,
-      "median",
-      @percentile_width,
-      "99th %"
-    ])
-    |> to_string
-  end
-
-  defp column_descriptors_for_memory(label_width) do
-    "\n~*s~*s~*s~*s~*s\n"
-    |> :io_lib.format([
-      -label_width,
-      "Name",
-      @average_width,
-      "average",
-      @deviation_width,
-      "deviation",
-      @median_width,
-      "median",
-      @percentile_width,
-      "99th %"
-    ])
-    |> to_string
-  end
-
-  defp column_descriptors_for_memory_same(label_width) do
-    "\n~*s~*s\n"
-    |> :io_lib.format([
-      -label_width,
-      "Name",
-      @average_width,
-      "Memory usage"
-    ])
-    |> to_string
-  end
-
-  defp label_width(scenarios) do
-    max_label_width =
-      scenarios
-      |> Enum.map(fn scenario -> String.length(scenario.name) end)
-      |> Stream.concat([@default_label_width])
-      |> Enum.max()
-
-    max_label_width + 1
-  end
-
-  @spec scenario_reports([Scenario.t()], unit_per_statistic, integer) :: [String.t()]
-  defp scenario_reports(scenarios, units, label_width) do
-    Enum.map(scenarios, fn scenario ->
-      format_scenario(scenario, units, label_width)
-    end)
-  end
-
-  defp scenario_reports_for_memory(scenarios, units, label_width) do
-    Enum.map(scenarios, fn scenario ->
-      format_scenario_for_memory(scenario, units, label_width)
-    end)
-  end
-
-  defp scenario_reports_for_memory_same(scenarios, units, label_width) do
-    Enum.map(scenarios, fn scenario ->
-      format_scenario_for_memory_same(scenario, units, label_width)
-    end)
-  end
-
-  @spec format_scenario(Scenario.t(), unit_per_statistic, integer) :: String.t()
-  defp format_scenario(
-         %Scenario{
-           name: name,
-           run_time_statistics: %Statistics{
-             average: average,
-             ips: ips,
-             std_dev_ratio: std_dev_ratio,
-             median: median,
-             percentiles: %{99 => percentile_99}
-           }
-         },
-         %{run_time: run_time_unit, ips: ips_unit},
-         label_width
-       ) do
-    "~*s~*ts~*ts~*ts~*ts~*ts\n"
-    |> :io_lib.format([
-      -label_width,
-      name,
-      @ips_width,
-      ips_out(ips, ips_unit),
-      @average_width,
-      run_time_out(average, run_time_unit),
-      @deviation_width,
-      deviation_out(std_dev_ratio),
-      @median_width,
-      run_time_out(median, run_time_unit),
-      @percentile_width,
-      run_time_out(percentile_99, run_time_unit)
-    ])
-    |> to_string
-  end
-
-  defp format_scenario_for_memory(
-         %Scenario{
-           name: name,
-           memory_usage_statistics: %Statistics{
-             average: average,
-             std_dev_ratio: std_dev_ratio,
-             median: median,
-             percentiles: %{99 => percentile_99}
-           }
-         },
-         %{memory: memory_unit, ninety_ninth: ninety_ninth},
-         label_width
-       ) do
-    "~*ts~*ts~*ts~*ts~*ts\n"
-    |> :io_lib.format([
-      -label_width,
-      name,
-      @average_width,
-      run_time_out(average, memory_unit),
-      @deviation_width,
-      deviation_out(std_dev_ratio),
-      @median_width,
-      run_time_out(median, memory_unit),
-      @percentile_width,
-      run_time_out(percentile_99, ninety_ninth)
-    ])
-    |> to_string
-  end
-
-  defp format_scenario_for_memory_same(
-         %Scenario{
-           name: name,
-           memory_usage_statistics: %Statistics{
-             average: average
-           }
-         },
-         %{memory: memory_unit},
-         label_width
-       ) do
-    "~*ts~*ts\n"
-    |> :io_lib.format([
-      -label_width,
-      name,
-      @average_width,
-      run_time_out(average, memory_unit)
-    ])
-    |> to_string
-  end
-
-  defp ips_out(ips, unit) do
-    Count.format({Count.scale(ips, unit), unit})
-  end
-
-  defp run_time_out(run_time, unit) do
-    Duration.format({Duration.scale(run_time, unit), unit})
-  end
-
-  defp deviation_out(std_dev_ratio) do
-    DeviationPercent.format(std_dev_ratio)
-  end
-
-  @spec comparison_report([Scenario.t()], unit_per_statistic, integer, map) :: [String.t()]
-  defp comparison_report(scenarios, units, label_width, config)
-
-  defp comparison_report([_scenario], _, _, _) do
-    # No need for a comparison when only one benchmark was run
-    []
-  end
-
-  defp comparison_report(_, _, _, %{comparison: false}) do
-    []
-  end
-
-  defp comparison_report([scenario | other_scenarios], units, label_width, _) do
-    [
-      descriptor("Comparison"),
-      reference_report(scenario, units, label_width)
-      | comparisons(scenario, units, label_width, other_scenarios)
-    ]
-  end
-
-  defp comparison_report_for_memory(scenarios, units, label_width, config)
-
-  defp comparison_report_for_memory([_scenario], _, _, _) do
-    # No need for a comparison when only one benchmark was run
-    []
-  end
-
-  defp comparison_report_for_memory(_, _, _, %{comparison: false}) do
-    []
-  end
-
-  defp comparison_report_for_memory([scenario | other_scenarios], units, label_width, _) do
-    [
-      descriptor("Comparison"),
-      reference_report_for_memory(scenario, units, label_width)
-      | comparisons_for_memory(scenario, units, label_width, other_scenarios)
-    ]
-  end
-
-  defp reference_report(
-         %Scenario{
-           name: name,
-           run_time_statistics: %Statistics{ips: ips}
-         },
-         %{ips: ips_unit},
-         label_width
-       ) do
-    "~*s~*s\n"
-    |> :io_lib.format([-label_width, name, @ips_width, ips_out(ips, ips_unit)])
-    |> to_string
-  end
-
-  defp reference_report_for_memory(
-         %Scenario{
-           name: name,
-           memory_usage_statistics: %Statistics{median: median}
-         },
-         %{memory: memory_unit},
-         label_width
-       ) do
-    "~*s~*s\n"
-    |> :io_lib.format([-label_width, name, @median_width, run_time_out(median, memory_unit)])
-    |> to_string
-  end
-
-  @spec comparisons(Scenario.t(), unit_per_statistic, integer, [Scenario.t()]) :: [String.t()]
-  defp comparisons(
-         %Scenario{run_time_statistics: reference_stats},
-         units,
-         label_width,
-         scenarios_to_compare
-       ) do
-    Enum.map(scenarios_to_compare, fn scenario = %Scenario{run_time_statistics: job_stats} ->
-      slower = reference_stats.ips / job_stats.ips
-      format_comparison(scenario, units, label_width, slower)
-    end)
-  end
-
-  defp comparisons_for_memory(
-         %Scenario{memory_usage_statistics: reference_stats},
-         units,
-         label_width,
-         scenarios_to_compare
-       ) do
-    Enum.map(scenarios_to_compare, fn scenario = %Scenario{memory_usage_statistics: job_stats} ->
-      slower =
-        if job_stats.median == 0 do
-          0.0
-        else
-          reference_stats.median / job_stats.median
-        end
-
-      format_comparison_for_memory(scenario, units, label_width, slower)
-    end)
-  end
-
-  defp format_comparison(
-         %Scenario{name: name, run_time_statistics: %Statistics{ips: ips}},
-         %{ips: ips_unit},
-         label_width,
-         slower
-       ) do
-    ips_format = ips_out(ips, ips_unit)
-
-    "~*s~*s - ~.2fx slower\n"
-    |> :io_lib.format([-label_width, name, @ips_width, ips_format, slower])
-    |> to_string
-  end
-
-  defp format_comparison_for_memory(
-         %Scenario{name: name, memory_usage_statistics: %Statistics{median: median}},
-         %{memory: memory_unit},
-         label_width,
-         slower
-       ) do
-    median_format = run_time_out(median, memory_unit)
-
-    "~*s~*s - ~.2fx more memory\n"
-    |> :io_lib.format([-label_width, name, @median_width, median_format, slower])
-    |> to_string
-  end
-
-  @spec descriptor(String.t()) :: String.t()
-  defp descriptor(header_str) do
-    "\n#{header_str}: \n"
-  end
+  defp input_header(input) when input == @no_input_marker, do: ""
+  defp input_header(input), do: "\n##### With input #{input} #####"
 end
