+++ lib/benchee/formatters/console/run_time.ex
@@ -1,0 +1,241 @@
+defmodule Benchee.Formatters.Console.RunTime do
+  @moduledoc """
+  This deals with just the formatting of the run time results. They are similar
+  to the way the memory results are formatted, but different enough to where the
+  abstractions start to break down pretty significantly, so I wanted to extract
+  these two things into separate modules to avoid confusion.
+  """
+
+  alias Benchee.{Statistics, Benchmark.Scenario, Conversion, Formatters.Console.Helpers}
+  alias Benchee.Conversion.{Count, Unit}
+
+  @type unit_per_statistic :: %{atom => Unit.t()}
+
+  @ips_width 13
+  @average_width 15
+  @deviation_width 11
+  @median_width 15
+  @percentile_width 15
+  @minimum_width 15
+  @maximum_width 15
+  @sample_size_width 15
+  @mode_width 25
+
+  @doc """
+  Formats the run time statistics to a report suitable for output on the CLI.
+
+  ## Examples
+
+  ```
+  iex> scenarios = [
+  ...>   %Benchee.Benchmark.Scenario{
+  ...>     name: "My Job", run_time_statistics: %Benchee.Statistics{
+  ...>       average: 200.0, ips: 5000.0,std_dev_ratio: 0.1, median: 190.0, percentiles: %{99 => 300.1},
+  ...>       minimum: 100.1, maximum: 200.2, sample_size: 10_101, mode: 333.2
+  ...>     }
+  ...>   },
+  ...>   %Benchee.Benchmark.Scenario{
+  ...>     name: "Job 2", run_time_statistics: %Benchee.Statistics{
+  ...>       average: 400.0, ips: 2500.0, std_dev_ratio: 0.2, median: 390.0, percentiles: %{99 => 500.1},
+  ...>       minimum: 200.2, maximum: 400.4, sample_size: 20_202, mode: [612.3, 554.1]
+  ...>     }
+  ...>   }
+  ...> ]
+  iex> configuration = %{comparison: false, unit_scaling: :best, extended_statistics: true}
+  iex> Benchee.Formatters.Console.RunTime.format_scenarios(scenarios, configuration)
+  ["\nName             ips        average  deviation         median         99th %\n",
+  "My Job           5 K         200 μs    ±10.00%         190 μs      300.10 μs\n",
+  "Job 2         2.50 K         400 μs    ±20.00%         390 μs      500.10 μs\n",
+  "\nExtended statistics: \n",
+  "\nName           minimum        maximum    sample size                     mode\n",
+  "My Job       100.10 μs      200.20 μs        10.10 K                333.20 μs\n",
+  "Job 2        200.20 μs      400.40 μs        20.20 K     612.30 μs, 554.10 μs\n"]
+
+  ```
+
+  """
+  @spec format_scenarios([Scenario.t()], map) :: [String.t(), ...]
+  def format_scenarios(scenarios, config) do
+    %{unit_scaling: scaling_strategy} = config
+    units = Conversion.units(scenarios, scaling_strategy, :run_time)
+    label_width = Helpers.label_width(scenarios)
+
+    [
+      column_descriptors(label_width)
+      | scenario_reports(scenarios, units, label_width) ++
+          comparison_report(scenarios, units, label_width, config) ++
+          extended_statistics_report(scenarios, units, label_width, config)
+    ]
+  end
+
+  @spec extended_statistics_report([Scenario.t()], unit_per_statistic, integer, map) :: [
+          String.t()
+        ]
+  defp extended_statistics_report(_, _, _, %{extended_statistics: false}) do
+    []
+  end
+
+  defp extended_statistics_report(scenarios, units, label_width, _config) do
+    [
+      Helpers.descriptor("Extended statistics"),
+      extended_column_descriptors(label_width)
+      | extended_statistics(scenarios, units, label_width)
+    ]
+  end
+
+  @spec extended_statistics([Scenario.t()], unit_per_statistic, integer) :: [String.t()]
+  defp extended_statistics(scenarios, units, label_width) do
+    Enum.map(scenarios, fn scenario ->
+      format_scenario_extended(scenario, units, label_width)
+    end)
+  end
+
+  @spec format_scenario_extended(Scenario.t(), unit_per_statistic, integer) :: String.t()
+  defp format_scenario_extended(scenario, %{run_time: run_time_unit}, label_width) do
+    %Scenario{
+      name: name,
+      run_time_statistics: %Statistics{
+        minimum: minimum,
+        maximum: maximum,
+        sample_size: sample_size,
+        mode: mode
+      }
+    } = scenario
+
+    "~*s~*ts~*ts~*ts~*ts\n"
+    |> :io_lib.format([
+      -label_width,
+      name,
+      @minimum_width,
+      Helpers.run_time_out(minimum, run_time_unit),
+      @maximum_width,
+      Helpers.run_time_out(maximum, run_time_unit),
+      @sample_size_width,
+      Count.format(sample_size),
+      @mode_width,
+      Helpers.mode_out(mode, run_time_unit)
+    ])
+    |> to_string
+  end
+
+  @spec extended_column_descriptors(integer) :: String.t()
+  defp extended_column_descriptors(label_width) do
+    "\n~*s~*s~*s~*s~*s\n"
+    |> :io_lib.format([
+      -label_width,
+      "Name",
+      @minimum_width,
+      "minimum",
+      @maximum_width,
+      "maximum",
+      @sample_size_width,
+      "sample size",
+      @mode_width,
+      "mode"
+    ])
+    |> to_string
+  end
+
+  @spec column_descriptors(integer) :: String.t()
+  defp column_descriptors(label_width) do
+    "\n~*s~*s~*s~*s~*s~*s\n"
+    |> :io_lib.format([
+      -label_width,
+      "Name",
+      @ips_width,
+      "ips",
+      @average_width,
+      "average",
+      @deviation_width,
+      "deviation",
+      @median_width,
+      "median",
+      @percentile_width,
+      "99th %"
+    ])
+    |> to_string
+  end
+
+  @spec scenario_reports([Scenario.t()], unit_per_statistic, integer) :: [String.t()]
+  defp scenario_reports(scenarios, units, label_width) do
+    Enum.map(scenarios, fn scenario ->
+      format_scenario(scenario, units, label_width)
+    end)
+  end
+
+  @spec format_scenario(Scenario.t(), unit_per_statistic, integer) :: String.t()
+  defp format_scenario(scenario, %{run_time: run_time_unit, ips: ips_unit}, label_width) do
+    %Scenario{
+      name: name,
+      run_time_statistics: %Statistics{
+        average: average,
+        ips: ips,
+        std_dev_ratio: std_dev_ratio,
+        median: median,
+        percentiles: %{99 => percentile_99}
+      }
+    } = scenario
+
+    "~*s~*ts~*ts~*ts~*ts~*ts\n"
+    |> :io_lib.format([
+      -label_width,
+      name,
+      @ips_width,
+      Helpers.ips_out(ips, ips_unit),
+      @average_width,
+      Helpers.run_time_out(average, run_time_unit),
+      @deviation_width,
+      Helpers.deviation_out(std_dev_ratio),
+      @median_width,
+      Helpers.run_time_out(median, run_time_unit),
+      @percentile_width,
+      Helpers.run_time_out(percentile_99, run_time_unit)
+    ])
+    |> to_string
+  end
+
+  @spec comparison_report([Scenario.t()], unit_per_statistic, integer, map) :: [String.t()]
+  defp comparison_report(scenarios, units, label_width, config)
+
+  # No need for a comparison when only one benchmark was run
+  defp comparison_report([_scenario], _, _, _), do: []
+  defp comparison_report(_, _, _, %{comparison: false}), do: []
+
+  defp comparison_report([scenario | other_scenarios], units, label_width, _) do
+    [
+      Helpers.descriptor("Comparison"),
+      reference_report(scenario, units, label_width)
+      | comparisons(scenario, units, label_width, other_scenarios)
+    ]
+  end
+
+  defp reference_report(scenario, %{ips: ips_unit}, label_width) do
+    %Scenario{
+      name: name,
+      run_time_statistics: %Statistics{ips: ips}
+    } = scenario
+
+    "~*s~*s\n"
+    |> :io_lib.format([-label_width, name, @ips_width, Helpers.ips_out(ips, ips_unit)])
+    |> to_string
+  end
+
+  @spec comparisons(Scenario.t(), unit_per_statistic, integer, [Scenario.t()]) :: [String.t()]
+  defp comparisons(scenario, units, label_width, scenarios_to_compare) do
+    %Scenario{run_time_statistics: reference_stats} = scenario
+
+    Enum.map(scenarios_to_compare, fn scenario = %Scenario{run_time_statistics: job_stats} ->
+      slower = reference_stats.ips / job_stats.ips
+      format_comparison(scenario, units, label_width, slower)
+    end)
+  end
+
+  defp format_comparison(scenario, %{ips: ips_unit}, label_width, slower) do
+    %Scenario{name: name, run_time_statistics: %Statistics{ips: ips}} = scenario
+    ips_format = Helpers.ips_out(ips, ips_unit)
+
+    "~*s~*s - ~.2fx slower\n"
+    |> :io_lib.format([-label_width, name, @ips_width, ips_format, slower])
+    |> to_string
+  end
+end
